+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Add handled response type to profiler item
+- Handle Impure Decoder
+  - Cache response before Impure Decoder decodes the response
+  - Fix scope of the fetch query/send command methods, which may overlaps with nested queries in impure decoding
+- Allow setting verbosity for a profiler
 
 ## 1.2.1 - 2022-03-09
 - Fix profiling last used decoders when there are none